+++ test/application_system_test_case.rb
@@ -63,4 +63,9 @@ class ApplicationSystemTestCase < ActionDispatch::SystemTestCase
   def go_to_dashboard
     click_link "DARIA - #{(ENV['FUND'] ? ENV['FUND'] : Rails.env)}"
   end
+
+  def click_away_from_field
+    find('body').click
+    wait_for_ajax
+  end
 end
